Added Scaling, moved FishImages to its own file
